+++ src/Validators/ResourceValidator.php
@@ -48,7 +48,7 @@ class ResourceValidator implements ResourceValidatorInterface
     private $expectedType;
 
     /**
-     * @var int|null|string
+     * @var string|null
      */
     private $expectedId;
 
@@ -87,7 +87,7 @@ class ResourceValidator implements ResourceValidatorInterface
     ) {
         $this->errorFactory = $errorFactory;
         $this->expectedType = $expectedType;
-        $this->expectedId = $expectedId;
+        $this->expectedId = $expectedId ? (string) $expectedId : null;
         $this->attributes = $attributes;
         $this->relationships = $relationships;
         $this->context = $context;
@@ -148,8 +148,26 @@ class ResourceValidator implements ResourceValidatorInterface
 
         $type = $resource->get(ResourceObjectInterface::TYPE);
 
+        /** Type must be string */
+        if (!is_string($type)) {
+            $this->addError($this->errorFactory->memberStringExpected(
+                $resource::TYPE,
+                P::dataType()
+            ));
+            return false;
+        }
+
+        /** Type must not be empty */
+        if (empty($type)) {
+            $this->addError($this->errorFactory->memberEmptyNotAllowed(
+                $resource::TYPE,
+                P::dataType()
+            ));
+            return false;
+        }
+
         /** Must be the expected type */
-        if (empty($type) || !$this->isSupportedType($type)) {
+        if (!$this->isSupportedType($type)) {
             $this->addError($this->errorFactory->resourceUnsupportedType($this->expectedType, $type));
             return false;
         }
@@ -163,14 +181,39 @@ class ResourceValidator implements ResourceValidatorInterface
      */
     protected function validateId(ResourceObjectInterface $resource)
     {
+        /** If we are not expecting an id, and one has not been provided, we can return true. */
+        if (!$this->isExpectingId() && !$resource->has($resource::ID)) {
+            return true;
+        }
+
         /** If expecting an id, one must be provided */
-        if (!is_null($this->expectedId) && !$resource->has(ResourceObjectInterface::ID)) {
+        if (!$resource->has(ResourceObjectInterface::ID)) {
             $this->addError($this->errorFactory->memberRequired(ResourceObjectInterface::ID, P::data()));
             return false;
         }
 
+        $id = $resource->get($resource::ID);
+
+        /** Id must be a string */
+        if (!is_string($id)) {
+            $this->addError($this->errorFactory->memberStringExpected(
+                $resource::ID,
+                P::dataId()
+            ));
+            return false;
+        }
+
+        /** Id must not be empty */
+        if (empty($id)) {
+            $this->addError($this->errorFactory->memberEmptyNotAllowed(
+               $resource::ID,
+               P::dataId()
+            ));
+            return false;
+        }
+
         /** If expecting an id, must match the one we're expecting */
-        if (!is_null($this->expectedId) && $this->expectedId != $resource->getId()) {
+        if ($this->isExpectingId() && $this->expectedId != $resource->getId()) {
             $this->addError($this->errorFactory->resourceUnsupportedId(
                 $this->expectedId,
                 $resource->getId()
@@ -262,4 +305,13 @@ class ResourceValidator implements ResourceValidatorInterface
 
         return false;
     }
+
+    /**
+     * @return bool
+     */
+    protected function isExpectingId()
+    {
+        return !is_null($this->expectedId);
+    }
+
 }